+++ backend/Domain/Models/Client.cs
@@ -4,10 +4,10 @@ namespace Domain.Models;
 
 public class Client : ITableDataBase
 {
-    public int? Id { get; set; }
+    public long? Id { get; set; }
     public string Ip { get; set; }
     
-    public Client(int? id, string ip)
+    public Client(long? id, string ip)
     {
         Id = id;
         Ip = ip;
