+++ backend/Domain/Models/Owner.cs
@@ -7,11 +7,11 @@ namespace Domain.Models;
 
 public class Owner : TgAccount, ITableDataBase, ITakeRelatedData
 {
-    [JsonPropertyName("restaurant_ids")] [NotMapped] public List<int>? RestaurantIds { get; set; }
+    [JsonPropertyName("restaurant_ids")] [NotMapped] public List<long>? RestaurantIds { get; set; }
 
     public Owner(){}
-    public Owner(int? id, string username) : base(id, username){}
-    public Owner(int? id, string username, List<int>? restaurantIds) : base(id, username)
+    public Owner(long? id, string username) : base(id, username){}
+    public Owner(long? id, string username, List<long>? restaurantIds) : base(id, username)
     {
         RestaurantIds = restaurantIds;
     }
