+++ backend/Domain/Models/Table.cs
@@ -6,14 +6,14 @@ namespace Domain.Models;
 
 public class Table : ITableDataBase
 {
-    public int? Id { get; set; }
-    [JsonPropertyName("restaurant_id")] public int RestaurantId { get; set; }
+    public long? Id { get; set; }
+    [JsonPropertyName("restaurant_id")] public long RestaurantId { get; set; }
     public int Number { get; set; }
     public string Token { get; set; }
     [JsonPropertyName("qr_code_image_path")] public string QRCodeImagePath { get; set; }
 
     public Table(){}
-    public Table(int? id, int restaurantId, int number)
+    public Table(long? id, long restaurantId, int number)
     {
         Id = id;
         RestaurantId = restaurantId;
