+++ backend/Domain/Models/TgAccount.cs
@@ -2,11 +2,11 @@ namespace Domain.Models;
 
 public class TgAccount
 {
-    public int? Id { get; set; }
+    public long? Id { get; set; }
     public string Username { get; set; }
 
     protected TgAccount(){}
-    protected TgAccount(int? id, string username)
+    protected TgAccount(long? id, string username)
     {
         Id = id;
         Username = username;
